Refactors Adapters: * They're now final endpoints. * They don't need to call `@app.call` anymore. * They don't inherit from `Faraday::Middleware` anymore.
